LEKSE-61 Refactor interfaces by using index file
