+++ src/interfaces/index.ts
@@ -2,5 +2,6 @@ import ISubject from './ISubject';
 import IGrade from './IGrade';
 import IQuestion from './IQuestion';
 import IStatus from './IStatus';
+import IMessage from './IMessage';
 
-export { ISubject, IQuestion, IGrade, IStatus };
+export { ISubject, IQuestion, IGrade, IStatus, IMessage };
